raffle update to select from checkins table and send request to Onesignal
This expects that the checkins table is being populated which I believe is not yet in place.
For testing purposes, manually populate this table.

Also, the mobile app needs to be updated to send a user tag with the users email address. I have a this stubbed out in home.ts in the mobile app. To test, uncomment the sendTag lines in the mobile app and use the matching email address that is in the checkins table.

Words and stuff.